feat: ads are now cached and prefetched
The flow of the Ads Controller is now working in a manner that is friendlier to the way a UITableView would work. Ads are fetched using a fetchAds method and cached in each Ads Controller. The views may then be built out appropriately from cached ads. Only stops that have not yet gottan any ads are requested for.

Potentially we might want to cache requested stop ads that had no ads as well to avoid adding new ads where none existed before.